Separate concerns add all querries into db file [Chore-#164871791]
