+++ database/db.py
@@ -2,6 +2,16 @@ import psycopg2
 from psycopg2.extras import RealDictCursor
 from instance.config import runtime_mode
 import os
+from datetime import date
+from werkzeug.security import generate_password_hash, check_password_hash
+from api.utilitiez.auth_token import get_current_identity
+from api.utilitiez.responses import (
+    duplicate_subject,
+    duplicate_message,
+)
+from api.utilitiez.responses import (
+    duplicate_email,
+)
 
 
 class DatabaseConnection:
@@ -76,7 +86,169 @@ class DatabaseConnection:
             """Function for connecting to appropriate database"""
             return psycopg2.connect(dbname='postgres', user='postgres', host='localhost', password='bekeplar')
 
-        
+    
+    def insert_user(self, **kwargs):
+        """User class method for adding new user to the users database"""
+        first_name = kwargs["first_name"]
+        last_name = kwargs["last_name"]
+        email = kwargs["email"]
+        user_password = generate_password_hash(kwargs["password"])
+
+        # Querry for adding a new user into users_db
+        sql = (
+            "INSERT INTO users ("
+            "first_name,"
+            "last_name, "
+            "email,"
+            "user_password )VALUES ("
+            f"'{first_name}', '{last_name}','{email}',"
+            f"'{user_password}') returning "
+            "user_id, first_name as firstname,"
+            "last_name as lastname,"
+            "email as email"
+        )
+        self.cursor_database.execute(sql)
+        new_user = self.cursor_database.fetchone()
+        return new_user
+
+    def check_if_user_exists(self, email):
+        """Making sure that a user's email is unique"""
+        user_exists_sql = (
+            "SELECT email from users where "
+            f" email='{email}';"
+        )
+        self.cursor_database.execute(user_exists_sql)
+        user_exists = self.cursor_database.fetchone()
+        error = {}
+
+        if user_exists and user_exists.get("email") == email:
+            error["email"] = duplicate_email
+
+        return error
+
+    def is_valid_credentials(self, email, user_password):
+        """Function for verrifying user credentials before login"""
+        sql = (
+            "SELECT user_id, email ,user_password FROM users \
+                where email="
+            f"'{email}';"
+        )
+        self.cursor_database.execute(sql)
+
+        user_details = self.cursor_database.fetchone()
+
+        if (
+                user_details
+                and user_details.get("email") == email
+                and check_password_hash(
+                user_details.get("user_password"), user_password
+        )
+        ):
+            id = user_details.get("user_id")
+
+            return id
+        return None
+    
+    def create_message(self, **kwargs):
+        """Function for adding a new message to the database"""
+        subject = kwargs.get("subject")
+        message = kwargs.get("message")
+        sender_status = "sent"
+        receiver_status = "unread"
+        receiver_id = kwargs.get("receiver_id")
+        sender_id = kwargs.get("user_id")
+        parent_message_id = kwargs.get("parent_message_id")
+        created_on = date.today()
+
+        # sql command for inserting a new message in the database
+        sql = (
+            "INSERT INTO messages ("
+            "subject, message, sender_status, receiver_status, receiver_id, sender_id, parent_message_id, created_on"
+            ")VALUES ("
+            f"'{subject}', '{message}','{sender_status}', '{receiver_status}',"
+            f"'{sender_id}', '{receiver_id}', '{parent_message_id}' ,'{created_on}') returning "
+            "message_id,subject as subject,"
+            "message as message, "
+            "sender_status as sender_status,"
+            "receiver_status as receiver_status, "
+            "receiver_id as receiver_id, "
+            "parent_message_id as parent_message_id, "
+            "created_on as created_on, "
+            "sender_id as sender_id;"
+        )
+        self.cursor_database.execute(sql)
+        new_incident = self.cursor_database.fetchone()
+        return new_incident
+
+
+    def check_duplicate_message(self, subject, message):
+        """Testing for uniqueness of a message."""
+        exists_query = (
+            "SELECT subject, Message from messages where "
+            f"subject ='{subject}' OR message='{message}';"
+        )
+        self.cursor_database.execute(exists_query)
+        message_exists = self.cursor_database.fetchone()
+        error = {}
+        if message_exists and message_exists.get("subject") == subject:
+            error["subject"] = duplicate_subject
+
+        if message_exists and message_exists.get("message") == message:
+            error["message"] = duplicate_message
+        return error
+
+
+    def get_message_record(self, msg_id, owner_id):
+        """Method to return a given message by id"""
+        sql = (
+            f"SELECT * FROM messages WHERE message_id='{msg_id}' \
+                    AND receiver_id='{owner_id}';"
+        )
+        self.cursor_database.execute(sql)
+        return self.cursor_database.fetchone()
+
+
+    def get_inbox_record(self, owner_id, msg_id):
+        """Method to delete a given message from user inbox by id."""
+        sql = (
+            f"SELECT * FROM messages WHERE receiver_id='{owner_id}' \
+                    AND message_id='{msg_id}';"
+        )
+        self.cursor_database.execute(sql)
+        return self.cursor_database.fetchone()
+
+
+    def get_sent_messages(self, owner_id):
+        """Function which returns all sent messages by a user."""
+        sql = (
+            f"SELECT * FROM messages WHERE sender_id='{owner_id}';"
+        )
+        self.cursor_database.execute(sql)
+        return self.cursor_database.fetchone()
+
+
+    def get_all_received_messages(self, owner_id):
+        """Function for getting all received messages."""
+        sql = (
+            f"SELECT * FROM messages WHERE receiver_id='{owner_id}';"
+        )
+        self.cursor_database.execute(sql)
+        return self.cursor_database.fetchone()
+
+
+    def delete_inbox_mail(self, msg_id, user_id):
+        """Function to delete a user's inbox mail."""
+        sql = (
+            f"DELETE FROM messages WHERE receiver_id='{user_id}' "
+            f"AND message_id='{msg_id}' returning *;"
+        )
+        self.cursor_database.execute(sql)
+        return self.cursor_database.fetchone()
+
+
+
+ 
+
     def drop_table(self, table_name):
             """
             Drop tables after tests
